fix: CommunityItem 수정
