+++ src/components/CardList/CardList.tsx
@@ -2,10 +2,10 @@ import { useContext, useEffect, useState } from 'react';
 
 import axios from 'axios';
 import { HomeCardListContext } from 'contexts/homeCardMenuContext';
-import { useLocation, useMatch } from 'react-router-dom';
+// import { useLocation, useMatch } from 'react-router-dom';
 
 import PostCard from 'components/commons/PostCard/PostCard';
-import CommunityList from 'components/CommunityList/CommunityList';
+import CommunityItem from 'components/CommunityItem/CommunityItem';
 
 import * as S from './CardList.style';
 
@@ -25,15 +25,56 @@ export const CardList = () => {
     crawlingData();
   }, [homeCardList?.getParams]);
 
-  const location = useLocation();
-  const CommunityPathMatch = useMatch('/?menu=community');
-  console.log(location.search, CommunityPathMatch);
+  console.log(data);
+  // const location = useLocation();
+  // const CommunityPathMatch = useMatch('/?menu=community');
 
   return (
     <S.CardListWrapper>
       {location.search.includes('community') ? (
         <section>
-          {data.map((el, idx) => {
+          {data &&
+            data.map((el, idx) => {
+              return (
+                <CommunityItem
+                  key={idx}
+                  user={el.user}
+                  dateCreation={el.createAt}
+                  title={el.title}
+                  path={el.path}
+                  viewCount={el.view}
+                  likeCount='1234'
+                  scrapCount='4234'
+                />
+              );
+            })}
+        </section>
+      ) : (
+        <S.Section>
+          {data &&
+            data?.map((el, idx) => {
+              return (
+                <PostCard
+                  key={idx}
+                  mainImage={el.mainImage}
+                  infoHost={el.enterprise}
+                  title={el.title}
+                  dateDday={el.Dday}
+                  dateCreation={el.applicationPeriod}
+                  scrapCount='4234'
+                  viewCount={el.view}
+                  location='111'
+                  isPickButton
+                  isPostCardTag
+                  index={idx}
+                />
+              );
+            })}
+        </S.Section>
+      )}
+      {/* <section>
+        {data &&
+          data.map((el, idx) => {
             return (
               <CommunityList
                 key={idx}
@@ -47,44 +88,6 @@ export const CardList = () => {
               />
             );
           })}
-        </section>
-      ) : (
-        <S.Section>
-          {data.map((el, idx) => {
-            return (
-              <PostCard
-                key={idx}
-                mainImage={el.mainImage}
-                infoHost={el.enterprise}
-                title={el.title}
-                dateDday={el.Dday}
-                dateCreation={el.applicationPeriod}
-                scrapCount='4234'
-                viewCount={el.view}
-                location='111'
-                isPickButton
-                isPostCardTag
-                index={idx}
-              />
-            );
-          })}
-        </S.Section>
-      )}
-      {/* <section>
-        {data.map((el, idx) => {
-          return (
-            <CommunityList
-              key={idx}
-              user={el.user}
-              dateCreation={el.createAt}
-              title={el.title}
-              path={el.path}
-              viewCount={el.view}
-              likeCount='1234'
-              scrapCount='4234'
-            />
-          );
-        })}
       </section> */}
     </S.CardListWrapper>
   );
