+++ src/components/CommunityItem/CommunityItem.styles.ts
